Init db file for tools
The repository should eventually become an exhaustive resource for
specification languages. Publications will always be fixed and
reliable references, but software projects tend to evolve or suffer
from bitrot, so tracking implementations may not be a feasible
ambition for tlzoo.

Note that Ioannis Filippidis (@johnyf) created and curates large lists
of software for formal methods:
https://github.com/johnyf/tool_lists